Rename lib Directory
