+++ lib/restiby/command.rb
