+++ lib/restiby/constants/commands.rb
