+++ lib/restiby/logger.rb
