+++ lib/restiby/multi_io.rb
