+++ lib/restiby/notifiers/base.rb
