+++ restiby.rb
@@ -9,17 +9,17 @@ require "optparse"
 require "ostruct"
 require "uri"
 require "yaml"
-require_relative "lib/restic/constants/commands"
-require_relative "lib/restic/multi_io"
-require_relative "lib/restic/logger"
-require_relative "lib/restic/location"
-require_relative "lib/restic/backend"
-require_relative "lib/restic/configuration"
-require_relative "lib/restic/command"
-require_relative "lib/restic/backup"
-require_relative "lib/restic/notifiers/base"
-require_relative "lib/restic/notifiers/discord"
-require_relative "lib/restic/notifiers/healthchecks_io"
-require_relative "lib/restic/parser"
+require_relative "lib/restiby/constants/commands"
+require_relative "lib/restiby/multi_io"
+require_relative "lib/restiby/logger"
+require_relative "lib/restiby/location"
+require_relative "lib/restiby/backend"
+require_relative "lib/restiby/configuration"
+require_relative "lib/restiby/command"
+require_relative "lib/restiby/backup"
+require_relative "lib/restiby/notifiers/base"
+require_relative "lib/restiby/notifiers/discord"
+require_relative "lib/restiby/notifiers/healthchecks_io"
+require_relative "lib/restiby/parser"
 
 Restiby::Parser.parse!
